Allow normal RW memory to be marked as executable
An assumption is made in the page table setup code that all RW memory must be
marked as XN. This is true for EL3 code. The logic to create page tables can be
reused to configure memory translation for secure ELs of lower privilege as
well. This assumption might not hold in these scenarios.

This patch reworks the logic to specify data and code access permission
attributes for normal memory. Code access permissions now comprise of two
bits. Data access permissions are still specified using a single bit. The two
types of permissions are ORed to form an aggregate value.

In this aggregate value when Bit[1] of code access permissions is unset, legacy
attributes are assumed and mapped as follows:

1. MT_RO i.e. Read-only normal memory is marked as executable
2. MT_RW i.e. Read-write normal memory is marked as non-executable

When Bit[1] of code access permissions is set, then Bit[0] is used to mark
memory as executable (Bit[0]=0) or non-executable (Bit[0]=1).

This approach enables backwards compatibility while allowing future code to
specify access permissions with greater flexibility.

Change-Id: I6641c49b4c75ca6c218cee2167c912a0725ff3b5