SFS/MM: Add PI v1.5 HOB library
The Platform Initialisation v1.5 specification defines a mechanism to hand over
information between the various UEFI boot phases. A HOB (Handover Block) is the
main data structure that is used to populate and ferry information. This
mechanism will be used to pass initialisation information to the MM secure
firmware service in subsequent patches. This patch adds a library (copied from
Hob library during the SEC phase in the ARM port in EDK2) and its dependencies
that enables creation and manipulation of a hob library in BL2.

Change-Id: I42f47937651d008688ba832e7237126048175af7