SFS/MM: HACK to make page table creation data driven
The library to create translation tables assumes that:

a. the memory for the tables will always be statically allocated within the
   library
b. only a single set of tables will be created in an image for the EL it is
   running in.

These assumptions prevent reuse of this library to create translation tables for
another exception level e.g. S-EL1. This is required to enable execution of
secure firmware services in S-EL0. This patch implements some rather nasty
workarounds to allow the caller of the core translation table creation functions
to specify a different memory map and memory for populating the tables.

Change-Id: I391d631054f3c181a1786b83bb052ee345e0eb52