SFS/MM: HACK to create MMU init. register context
The library to create translation tables exports the enable_mmu_elX() function
that programs the system registers associated with enabling the MMU. After
tables for another EL have been created, it is preferrable to instead copy the
values of these registers into the system register context that will be used for
entering the EL. This way any subsequent updates to the same registers can be
consolidated in the context before being applied at the time of ERETing into the
EL.

This patch hacks a new prepare_mmu_context_elX() function that populates the
values of the relevant system registers in memory provided by the caller. They
are subsequently used to enable the MMU for the target EL. This function is
defined only for EL1.

Change-Id: Ic7cf131568983c57af58ec849845d87d26d18fb7