SPCI ALP2+: Use SoC FW config to enable and jump into SEL2
This patch adds support to:

1. Determine if SEL2 is supported on the FVP platform
2. Enable SEL2 if present and specified in the SOC_FW_CONFIG FDT

BL31 reads the SoC configuration to determine if SEL2 must be enabled.

SPM reads the SCR_EL3 to determine if SEL2 has been enabled. It programs a jump
into SEL2 for the BL32 stage if this case. The behaviour of the SPM will be
quite different when SEL2 is enabled. This will be fleshed out in later patches.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>